✨(api) add a facet_sorting query string param to course search
We need facet sorting on course filters to be different depending
on the circumstances: in some cases we respect the default from
the filter definition, in other cases we may have good reason to
change it.

This is the case in our <SearchFilterGroupModal />. As we cannot
order eg. category search results by facet count, because the
search request for categories knows nothing about facets, we must
order the results alphabetically.

We're adding a query string param to accomodate this use case
without altering the default behavior.