♻️(frontend) improve modal handling of app element
Newer versions of react-modal include a new way to specify the
app root element to hide from the accessibility tree when the
modal is open.

This new way, using a prop instead of a static method call, lets
us get the root element more cleanly and avoid testing for a jest
constant in actual application code.