✨(frontend) add pagination to search filter facets search
It is often the case that facet categories have a lot of possible
filter values. We initially built the <SearchFilterGroupModal /> to
enable users to search through those values. However, we did not
set up pagination.

As it turns out, users want to be able to move down the list
to get to the value they want. We decided to implement infinite
scrolling using react-query and IntersectionObserver to make it
as effortless as possible for users.